Update Documentation and save vinal python file
